+++ app/service/biz/channels/internal/dao/dataobject/channel_message_boxe_do.go
@@ -1,10 +1,10 @@
 package dataobject
 
 type ChannelMessageBoxDO struct {
-	Id                  int32  `db:"id"`
-	SenderUserId        int32  `db:"sender_user_id"`
-	ChannelId           int32  `db:"channel_id"`
-	ChannelMessageBoxId int32  `db:"channel_message_box_id"`
+	Id                  int64  `db:"id"`
+	SenderUserId        int64  `db:"sender_user_id"`
+	ChannelId           int64  `db:"channel_id"`
+	ChannelMessageBoxId int64  `db:"channel_message_box_id"`
 	MessageId           int64  `db:"message_id"`
 	Date                int32  `db:"date"`
 	Deleted             int8   `db:"deleted"`
