+++ app/service/biz/channels/internal/dao/dataobject/channel_pts_update_do.go
@@ -2,7 +2,7 @@ package dataobject
 
 type ChannelPtsUpdateDO struct {
 	Id         int64  `db:"id"`
-	ChannelId  int32  `db:"channel_id"`
+	ChannelId  int64  `db:"channel_id"`
 	Pts        int32  `db:"pts"`
 	PtsCount   int32  `db:"pts_count"`
 	UpdateType int8   `db:"update_type"`
